Fix potential race where we're expecting a moniker to be attached
When we're creating a ResultSet for a symbol, we always attach
a moniker on it once it's created. We do that outside the main lock
against the dictionary for all symbols, so that way the creation
of the moniker isn't holding up operations for unrelated symbols.
This was fine because nothing cared about the moniker existing, until
we added additional support for inheritance, and the new spec now
needs us to link to monikers in some cases. There I put in an incorrect
assumption that the moniker might be there, which isn't the case
if a different thread is still making it.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1692879