refactor document outline to do ui updates on a background thread
refactored this in to a View/ViewModel pattern where the ViewModel _never_ needs to be on the UI thread and the View is asserted tp _always_ be. This is because the ViewModel implements INotifyPropertyChanged and uses XAML data-binding to the View. This results in WPF automatically marshaling data changes to the UI thread when necessary in a way that is more efficient that what we can do.

Added VisualStudioCodeWindowInfoService to handle getting VS data from on the UI thread when needed.