Simplify the handling of diagnostic item sources with missing references
Due to the asynchronous nature of how the workspace is populated,
it's possible for us to see a node representing an analyzer in the
solution explorer, but we can't put analyzers and generators underneath
it because it hasn't yet been added to the workspace. The logic we
had here was a bit convoluted: it was trying to handle the case where
the project might not have been added yet, but that case wasn't
correctly handled anyways: the type is only created with a ProjectId
already being passed in, which comes from the workspace that was
previously enumerated to get it. So that simply can't happen.
There was also a theoretical issue where if we ever changed the
workspace to remove multiple projects with a SolutionChanged event,
we would have leaked in that case.

Since we don't have to care about add-then-remove, we can just check
in th event handler for the project still existing. If anything
removes it, we can unsubscribe.