Continued analyzer assembly loading work (#66612)
* Continued analyzer assembly loading work

This change responds to several concerns raised in #66492:

- Collapse the `AnalyzerAssembyLoader` hierarchy by one level and move
  more logic into the base type.
- Fix `AssemblyLoadContext` in tests to not replicate types. That means
  we can safely pass `ITestOutputHelper` into the custom
  `AssemblyLoadContext`
- Further strengthen the test separation

I have deliberately not changed the file names to reflect the new type
names because I wanted the review process to be smooth. Once I have two
sign offs for the change I will send another commit which does the
following:

- Rename
    - DefaultAnalyzerAssemblyLoaderTests.cs -> AnalyzerAssemblyLoaderTests.cs
    - DefaultAnalyzerAssemblyLoader.Core.cs -> AnalyzerAssemblyLoader.Core.cs
    - DefaultAnalyzerAssemblyLoader.Desktop.cs -> AnalyzerAssemblyLoader.Desktop.cs
    - DefaultAnalyzerAssemblyLoader.cs -> AnalyzerAssemblyLoader.cs
- Move
    - `DefaultAnalyzerAssemblyLoader` into DefaultAnalyzerAssemblyLoader.cs
    - `InvokeUtil` to InvokeUtil.cs. As constructed every edit to this
      type causes Test Explorer to lose context on the theory branches.
      Doesn't come back until rebuild. Will be smoother editting in its own
      file.

That should get is into the style we expect here. After that I will
merge once tests are green.

* Unload AssemblyLoadContext after test completion

* Reuse AssemblyLoadTestFixture across test runs

This also has the added benefit that the entire .NET Core set of tests
run in less than one second. Previously it close to half a second per
test.

This will hopefully alleviate the pain for the runtime issue we're
hittnig running `AssemblyLoadContext` tests in CI. This reduces the
amount of compilations on this code path which seem to be the trigger
for the failure.

- https://github.com/dotnet/roslyn/issues/66621
- https://github.com/dotnet/runtime/issues/81108

* Apply suggestions from code review

Co-authored-by: Rikki Gibson <rikkigibson@gmail.com>

* PR feedback

* rename the files

* Fixup

---------

Co-authored-by: Rikki Gibson <rikkigibson@gmail.com>