Fix up local/non-local diagnostic classification in analyzer driver
Work towards #66968

**NOTE:** None of the changes in this PR affects analyzer execution in batch compile scenarios.

In order to ensure that we can minimize the execution scope for analyzers when computing diagnostics for a line span (lightbulb), we need to ensure that the local/non-local diagnostic classification in the analyzer driver is hardened. The analyzer driver already has logic to classify analyzer reported diagnostics into local and non-local diagnostics based on whether or not the reported diagnostic span intersects with the callback symbol/operation/node span. This change fixes up one case in this logic:
- If an analyzer registers an Operation/SyntaxNode action, but not within the scope of any start actions, i.e. OperationBlockStart/SymbolStart/CodeBlockStart, then any diagnostic reported outside the span of the callback operation/node should be considered a non-local diagnostic.

Adding the unit test identified one other minor issue, which is fixed with this PR:
- Ensure that we de-dupe the reported non-local diagnostics. We were already de-duping the reported local diagnostics.