Watch some places where we put reference assemblies on Unix
Generally, the language service is simply given a list of source
files or reference assemblies, and we have to watch all of them in case
any of them were to change. One problem we run into is we just get
files without any real sense of structure, and this can lead to
inefficiencies. For example, for reference assemblies, if we're given
a list of reference assemblies, we can naively add each file to a list
to watch, but at some point we either have to transition to a directory
watch for efficiency. And although we can do that, it's still time
being spent during a solution load when we can be doing better things.

One optimization we do on Windows is we know there are certain folders
where we're likely to get lots of references, but the folder is unlikely
to have many changes, so there's no real downside to just watching the
entire directory at once. For example, on Windows the .NET Framework
reference assemblies folder we can just stick a file watch on at the
very beginning, and then know we never have to watch any individual file
under that directory for the rest of the process. This extends that
optimization to Linux/Mac as well.