Disable analyzer telemetry for all but lightbulb invocation path
We enabled this analyer telemetry for all diagnostic computation code paths in 17.5, but this ended up generating a lot of traffic and we had to suppress the telemetry event. This change makes a targeted fix to disable the telemetry event for all paths except the lightbulb path. We want to collect this telemetry to help identify slow analyzers affecting lightbulb perf.

In future, we plan to revert this PR and re-enable the telemetry on all diagnostic computation paths using the new telemetry aggregation APIs exposed by the telemetry team.