Framework compilers package for SDK consumption (#66206)
* Framework compilers package for SDK consumption

This is the Roslyn side of the fix for [msbuild 7832](https://github.com/dotnet/msbuild/issues/7832)
It provides a NuPkg that contains _only_ the .NET Framework version of
the Roslyn compilers. This matches, almost identically, what is inserted
into the MSBuild install of Visual Studio.

The .NET SDK can now use this package to use a .NET Framework toolset
when building C# and VB projects. This will resolve the analyzer
inconsistency issues that come up when mixing .NET SDK and MSBuild
Framework toolsets at different versions in CI.

Related
- https://github.com/dotnet/roslyn/issues/61252
- https://github.com/dotnet/msbuild/issues/7832

The package verification

* PR feedback

* Full run only

* YAML we meet again

* round 2 YAML

* make the pain stop

* yaml variable expansion part 2

* yaml hates typos

* wohoo a code bug!

* it's always yaml

* Apply suggestions from code review

Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>

* readme

* can't type xml

* Exclude from source build

---------

Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>