Include explicitly BCL references for .NET 6 OOP
For BCL, we want to use the version provided by the runtime in VS, not the ones from the NuGet packages. This might not be possible if Roslyn is referencing a higher version than the one shipped as part of runtime in VS. For example, we could be referecing S.C.I 7.0 but VS still ships .NET 6. Usually this is a transient state, and the two would eventually be in sync. But we'd need to include those binaries during the out-of-sync period.