Adjust solution crawler logic to choose next best project to reduce expensive allocations
Fixes #22609

Background analysis in solution crawler currently chooses the documents/projects to analyze using the following priority order:
1. Active document
2. Open documents in a project that had one or more diagnostics in prior snapshot
3. Open documents on the remaining dependent projects of the active document project based on the project dependency graph
4. Rest of the open documents
5. Projects that had one or more diagnostics (If FSA is enabled)
6. Remaining dependent projects of the active document project based on the project dependency graph (If FSA is enabled)
7. Rest of the projects (If FSA is enabled)

The logic to compute if a project had any diagnostics in the prior snapshot has been known to be pretty expensive causing lot of allocations. Additionally, we are not sure if that is really
providing any additional user experience improvement for error list refresh. This change removes the checks in 2. and 6. above.