Merged PR 30834: Add component wrapper for telemetry
This changes telemetry to use a component wrapper, similar to intellicode. This ensures we don't need MEF to call to the component if it's available, and loads it into it's own ALC.

This also removes any dependencies the internal component has on Roslyn code, which allows it to exist in an isolated way.