Accept the service broker pipe name via LSP rather than a switch
By requiring the pipe name to be passed as a switch, we were requiring
that the Dev Kit server processes had to be running before we could
launch our process. This isn't strictly necessary; by moving to an LSP
notification we can connect to it asynchronously once the process is
running.