Merged PR 31357: Emulate soft-selection when filter text is empty
In the gif below, I typed `"` in both cases, the difference is the first time filter text is empty, the second time "s" was typed. You can see the selected item was not committed in the first case. This behavior is consistent with VS

The downside of this approach is there's no telling in UI whether the item would only be committed via TAB

![completion-selection.gif](https://dnceng.visualstudio.com/7ea9116e-9fac-403d-b258-b31fcf1bb293/_apis/git/repositories/7b863b8d-8cc3-431d-b06b-7136cc32bbe6/pullRequests/31357/attachments/completion-selection.gif)