Move LongPath check to MSBuild target (#68217)
The Roslyn repository effectively depends on having long paths enabled
at this point. Between NuGet and the VSSDK, both outside our control,
our repository runs right up against `MAX_PATH` limits even with a very
short user name. A user name of any significant length will quickly run
into issues in our repo.

The repository already warns about long path support, and has for some
time, but it does so in command line builds. That was a fine experience
when it was added as long path support was new and was only needed in a
few cases. Now though it's effectivtely needed for everyone so need to
give it more visibility by elevating it to an MSBuild warning.

This warning is _deliberately_ not configurable. Having long paths
disabled at this point means your inner loop dev experience is broken in
strange, hard to diagnose ways.

Co-authored-by: Jared Parsons <jared@popcornbear.org>