Ensure we only ask for a symbol's documentation comment once
Right now computing hover information is a fairly expensive portion
of an LSIF dump. One small hotspot was we were fetching documentation
comment information at least twice: once for the documentation itself
and once for computing the exception info. This ensures we fetch
it just once per generation.

This also generally cleans up the AbstractSymbolDescriptionBuilder:

1. Duplicate local functions for returns/value is unified.
2. Removes a bunch of specific logic in the builder which supported
   finding the correct documentation in favor of the helper we
   already had.

To support the second one, I added some methods to DocumentationComment
to return a DocumentationComment with the contents of a <param> tag
as the summary text for a specific parameter; this allows everything
else to be able to operate as if that's a regular summary tag and
removes some special cases in the code.