Don't rebind object initializer parents (#68449)
* Don't rebind object initializer parents

We had an issue where, when requesting semantic info for an expression nested in a collection or complex collection initializer, we'd attempt to bind the direct syntactic parent of that expression. That expression is never bindable, so it would always be rebound no matter what. This leads to particular issues for things like the LSIF indexer when encountering files with large (often generated) dictionary initializers.

* Simplify timeout logic

* Scope changes to just GetBindableParentNode

* Reduce number of iterations

* More amount fine-tuning

* Additional functional tests for complex initializer conversions.