Correct interpolated string addition parts counting (#68562)
* Correct interpolated string addition parts counting

We were counting the number of parts an interpolation had incorrectly for addition operators, counting the number of additions, not the number of interpolation components. This means we were calling `string.Concat` for such locations instead of using the `DefaultInterpolatedStringHandler` where we should have. Contributes to https://github.com/dotnet/runtime/issues/78991#issuecomment-1567524054.

* Restructure the code for better clarity.

* PR feedback.