Add ReportIVTs compiler mode (#68665)
This adds a new flag to the command-line compiler (and corresponding msbuild property) that reports on IVT information the compiler knows about and tries to help users figure out where IVT issues are coming from. Concretely, it changes the output in 3 ways:

1. After all regular diagnostics are calculated and emitted, we look for any diagnostics that are about inaccessible members. If we find any, and those members are from other assemblies, we print what assembly they're from.
2. After this, we print the current assembly identity, meaning name and public key.
3. Finally, we go through every reference passed to the compiler and print the assembly identity, wether the assembly has granted IVT to the assembly being compiled, and all IVTs the compiler sees this assembly as having.

Examples of what this output looks like are in the newly added commandline tests for the scenario.