Add effective severities custom property array to error log
Closes #67365

Add a new custom property array named `effectiveConfigurationLevels` to the `rule` section in sarif v2 error log output. If the severity of any analyzer diagnostic ID was configured via options (command line options, editorconfig, globalconfig, rulesets, etc.) for a part or whole of the compilation, we emit this array of effective severities.