Add failing test
The F1 service should produce `#if` instead of `if_CSharpKeyword` when
the cursor is on the `#if` symbol.

Add other tasks for the other preprocessor symbols related to
conditional compilation. The `#if` and `#else` fail, because the model
finds child nodes for `if` and `else`, respectively. The tests for
`#elif` and `#endif` pass, because the model does not find child nodes
for those tokens.