Update MockServiceProvider to opt out of return an SVsFeatureFlags
Right now the only place this is getting requested in tests is in
the option persisters; that code treats it as optional so we can just
return nothing rather than try to return a full mock. This is a targeted
fix for the 17.7 branch and I'll clean this up in main.