Update final test to passing state
I did not expect this test to be altered and start inserting a blank
line after the new using directive, but I also was not able to determine
the reason why this blank line appeared or resolve it without breaking
other tests. It seems like an edge case that would not be especially
problematic, but I still wanted to call out the change in its own
commit.