Fix analyzer conflicts on Linux (#69406)
This changes analyzer loading on Linux to be per user where before clashes could occur in the temp directory.

This fix initially started out by making our temp directory logic more robust. Essentially add per user sub directories, do extended permission checks, etc ... That exacerbated a number of existing corner cases in the code and forced us to put a lot more error handling into the workspaces layer.

After discussion with the IDE team I decided to take the approach of loading from a `Stream` on non-Windows platforms. The performance difference is neglible (1ms to 2ms per assembly) and it removes the overhead of copying them around on disk as well as the complexities that come with managing the shadow copy directory.

This will be observable to any analyzer that used `Assembly.Location` as it will now be `""`. That is only useful though for inspecting disks which is already illegal for analyzers hence it's considered an acceptable break.

To test the performance of this change I wrote a simple benchmark that loads assemblies directly from path, copy then load (simulate shadow copy) and loading them from stream. The results on Windows is pretty striking

|Method | Mean | Error | StdDev | Median|
-- | -- | -- | -- | --| 
|LoadViaPath | 1.552 ms | 0.0387 ms | 0.1142 ms | 1.529 ms|
|LoadViaPathShadow | 33.147 ms | 0.9000 ms | 2.6537 ms | 33.883 ms|
|LoadViaStream | 495.945 ms | 9.5040 ms | 9.7599 ms | 493.297 ms|

Notice the penalty of `LoadFromStream` is severe on Windows. That is due to the AV having to run on every load since it can't cache like it can when loading from path. On Linux though the results are much nicer

|Method | Mean | Error | StdDev | Median|
|-- | -- | -- | -- | --|
|LoadViaPath | 1.301 ms | 0.0258 ms | 0.0326 ms | 1.299 ms|
|LoadViaPathShadow | 4.126 ms | 0.3567 ms | 1.0060 ms | 3.742 ms|
|LoadViaStream | 2.514 ms | 0.0501 ms | 0.0536 ms | 2.497 m|

This means the change should be a small win on Linux but probably not in the noticable range. 

closes #65415