Avoid boxing when structs are used as collection keys. (#69443)
Change several structs to record structs such that they no longer are boxed when used as keys in MultiDictionary's internal ImmutableHashSet. When MultiDictionary's value is a struct and is created without specifying the value's IEqualityComparer, the default struct comparer is used. This equality comparer is notoriously slow due to boxing. By switching these structs to record structs, an implementation of IEquatable is generated and the default struct equality comparer isn't used.

Addresses https://github.com/dotnet/roslyn/issues/69165

Sam mentioned that there is an issue already logged in the roslyn-analyzers repo around adding an analyzer to detect this.

https://github.com/dotnet/roslyn-analyzers/issues/1640