Add support for reporting non-local LSP pull document diagnostics for open documents.
Fixes #5634

Implemented as per https://github.com/dotnet/vscode-csharp/issues/5634#issuecomment-1663178971

Currently, we only report local document diagnostics for open documents in the DocumentPullDiagnosticHandler. We report local and non-local diagnostics for closed documents, as well as no-location projects in the WorkspacePullDiagnosticHandler.

This change adds a new dynamically registered document diagnostic source in the public DocumentPullDiagnosticHandler with a new identifier to report non-local document diagnostics for open documents. I verified that after this change compilation end diagnostics are reported for open documents when FSA is enabled.