Remove use of Environment.CurrentDirectory (#69564)
The compiler should not make use of `Environment.CurrentDirectory` in
the libraries layer. That layer cannot depend on anything environment
related because doing so would break cases like VBCSCompiler which needs
to be environment agnostic.

The one use was in `AdditionalSourcesCollection` to verify that the
provided hint name was within the current working directory. After some
thought I decided the best course was to remove this restriction. The
ratoinale is once you actually thread through the current directory you
find that it must be added to the `Compilation`. Effectively everything
that hosts a `Compilation` must provide a working directory and that
goes against the goals of `Compilation` which seeks to be agnostic to
working directories.

Think allowing customers to specify the full paths was the lesser evil
here.

closes #69516