Split Microsoft.CodeAnalysis.Workspaces.MSBuild into two
This moves nearly all of the code that uses MSBuild to a separate
project, Microsoft.CodeAnalysis.Workspaces.MSBuild.BuildHost. This
new project is an executable project, and will eventually be the
separate process that is invoking MSBuild. For now, it's still
consumed as a regular project-to-project reference to keep everything
else working.

There are a few things that are still using MSBuild in the original
library, namely that we have some methods that take an ILogger and
the OpenSolutionAsync method calls into the solution file parsing APIs.
The method that takes an ILogger we will not be able to support in
an out-of-proc mode (although we can still maintain an in-proc compat
mode); the OpenSolutionAsync we will probably be able to address by
having the solution file parsing happening in-process since that
doesn't depend on any project file details.