Remove ProjectFileInfo.Log
The shape of this type was a bit deceiving: a ProjectFile has a Log
property, and this also having a Log property makes it appear like
there's one log for the basic loading of the project, and then a second
log for the actual loading for a specific TFM. This is however not the
case: ProjectFileInfo.Log would always be the same instance as the
ProjectFile it came from. Removing this is easy, as the only place
it was ever used was where we already had the ProjectFile in hand,
and it makes ProjectFileInfo a simple data type.