Include the BuildHost binaries in the Workspaces.MSBuild package
I'm taking this approach rather than creating another package since
this is only temporary. Once we move MSBuildWorkspace to directly
launch the BuildHost process (rather than loading it as assembly)
we'll need to deploy multiple content copies of the BuildHost since
we need to pick the right one at runtime.