Update the severity combo boxes in EditorConfig UX to reflect the correct semantics
Fixes #69176
Current UX and semantics for severity combo boxes in each of the tabs (Analyzers, Naming Style and Code Styles) in the EditorConfig UX
have Severity combo boxes with following strings: Disabled, Suggestion, Warning and Error. Disabled setting actually maps to "silent" severity value in editorconfig, which doesn't disable the analyzer, but converts it into a hidden analyzer for which code fixes are still offered, i.e. it is essentially a refactoring.
With this change, we make the following changes:
- Analyzers and Naming Styles tab now have following values for Severity combo box: Disabled, Refactoring Only, Suggestion, Warning and Error. Disabled now correctly maps to "none" in editorconfig, which disables the rule. Refactoring Only maps to "silent".
- Analyzers and Naming Styles tab has the following values for Severity combo box: Refactoring Only, Suggestion, Warning and Error. Disabled severity is not supported for code style settings. Code styles can instead be disabled by setting the option "Value" that turns off the style, such as "Yes/No", "Always/Never", etc. Theoretically, we can support the 'Disabled' severity, which translates to 'none' value in editorconfig. However, adding this support would require us to update all our IDE code style analyzers to turn themselves off when 'option.Notification.Severity is ReportDiagnostic.Suppress', which seems redundant.